Add enforcement option to date range picker defaults
The parameters of several methods (startDate, endDate, defaultToday, processDefault) in the DateRangePicker class have been updated to include an optional enforcement flag. If this flag is true and the state is null after hydration, the default state is enforced. This allows more control over whether the default date range should be applied when there is no set state.